backend/wallet: Refactor domain layer
