+++ backend/wallet/domain/src/repository/resource.rs
